Post Processing functionality implementation - Blur filter fixed - Bloom effect fixed - AntiAliasing effect fixed

+++ engine/src/resources/material.cpp
@@ -174,7 +174,7 @@ void Material::setTexture(const string &name, Texture *texture) {
     m_Textures[name] = texture;
 }
 
-uint8_t Material::surfaces() const {
+int32_t Material::surfaces() const {
     return m_Surfaces;
 }
